echo-notifications/requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
pkg-resources==0.0.0
requests==2.23.0
RPi.GPIO==0.7.0
urllib3==1.25.8
